+++ modules/Grades/includes/StudentAssignments.fnc.php
@@ -67,13 +67,29 @@ function StudentAssignmentSubmit( $assignment_id, &$error )
 	if ( isset( $_FILES[ 'submission_file' ] ) )
 	{
 		$file_attached_ext_white_list = array(
+			/**
+			 * Extensions white list.
+			 *
+			 * Common file types.
+			 * Obviously, we won't include executable types
+			 * .php, .sql, .js, .exe...
+			 * If you file type is not white listed,
+			 * put it in a ZIP archive!
+			 *
+			 * @link http://fileinfo.com/filetypes/common
+			 */
 			// Micro$oft Office.
 			'.doc',
 			'.docx',
 			'.xls',
 			'.xlsx',
+			'.xlr',
+			'.pps',
 			'.ppt',
 			'.pptx',
+			'.wps',
+			'.wpd',
+			'.rtf',
 			// Libre Office.
 			'.odt',
 			'.ods',
@@ -83,20 +99,49 @@ function StudentAssignmentSubmit( $assignment_id, &$error )
 			'.jpeg',
 			'.png',
 			'.gif',
-			// Sound.
+			'.bmp',
+			'.svg',
+			'.ico',
+			'.psd',
+			'.ai',
+			'.eps',
+			'.ps',
+			// Audio.
 			'.mp3',
 			'.ogg',
 			'.wav',
+			'.mid',
+			'.wma',
 			// Video.
 			'.avi',
 			'.mp4',
+			'.mpg',
 			'.ogv',
 			'.webm',
-			// Others.
-			'.zip',
+			'.wmv',
+			'.mov',
+			'.m4v',
+			'.flv',
+			'.swf',
+			// Text.
 			'.txt',
 			'.pdf',
+			'.md',
 			'.csv',
+			'.tex',
+			// Web.
+			'.xml',
+			'.xhtml',
+			'.html',
+			'.htm',
+			'.css',
+			'.rss',
+			// Compressed.
+			'.zip',
+			'.rar',
+			'.7z',
+			'.tar',
+			'.gz',
 		);
 
 		$student_name_RET = DBGet( DBQuery( "SELECT FIRST_NAME||' '||LAST_NAME AS NAME